+++ mm/page_alloc.c
@@ -5781,8 +5781,8 @@ void __init mem_init_print_info(const char *str)
 	 */
 #define adj_init_size(start, end, size, pos, adj) \
 	do { \
-		if (start <= pos && pos < end && size > adj) \
-			size -= adj; \
+		if ((unsigned long)(start) <= (unsigned long)(pos) && (unsigned long)(pos) < (unsigned long)(end) && (size) > (adj)) \
+			(size) -= (adj); \
 	} while (0)
 
 	adj_init_size(__init_begin, __init_end, init_data_size,
